Fixes the time before deletion and the form od the describe process.
