+++ core-server/src/main/java/org/orbisgis/orbisserver/coreserver/model/Result.java
@@ -38,7 +38,11 @@
  */
 package org.orbisgis.orbisserver.coreserver.model;
 
+import javax.xml.datatype.Duration;
 import javax.xml.datatype.XMLGregorianCalendar;
+import java.text.SimpleDateFormat;
+import java.util.Calendar;
+import java.util.Date;
 import java.util.List;
 
 /**
@@ -77,5 +81,20 @@ public class Result {
      */
     public void setOutputList(List<Output> outputList){
         this.outputList = outputList;
+    }
+
+    public String getRemainTime(){
+        Date date = new Date();
+        Date expDate = expirationDate.toGregorianCalendar().getTime();
+
+        SimpleDateFormat formatter = new SimpleDateFormat("MM/dd/yyyy hh:mm");
+        formatter.setTimeZone(expirationDate.toGregorianCalendar().getTimeZone());
+        long diff = expDate.getTime()-date.getTime();
+        diff /= 1000;
+        long days = diff/(60*60*24);
+        long hours = (diff-days*(60*60*24))/(60*60);
+        long minutes = (diff-days*(60*60*24)-hours*(60*60))/60;
+        String str = "" + days + "j" + hours + ":" + minutes;
+        return str;
     }
 }
